pump generation and plotting
