+++ src/perturbopy/postproc/__init__.py
@@ -17,7 +17,7 @@ from .calc_modes.dyna_pp import DynaPP
 from .dbs.units_dict import UnitsDict
 from .dbs.recip_pt_db import RecipPtDB
 
-from .utils import constants, plot_tools, lattice, spectra, timing
+from .utils import constants, plot_tools, lattice, spectra_generate_pulse, timing
 
 import warnings
 import os
